Move horologist.audio tests to correct place
