+++ audio/src/androidTest/java/com/google/android/horologist/audio/AudioOutputRepositoryTest.kt
@@ -14,12 +14,9 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.compose.layout
+package com.google.android.horologist.audio
 
 import androidx.test.platform.app.InstrumentationRegistry
-import com.google.android.horologist.audio.AudioOutput
-import com.google.android.horologist.audio.ExperimentalAudioApi
-import com.google.android.horologist.audio.SystemAudioOutputRepository
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
